requirements.txt
autopep8==2.0.1
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.10.0
pyee==9.0.4
python-dotenv==0.21.0
slack-bolt==1.16.0
slack-sdk==3.19.5
tomli==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
